+++ src/main/java/lotto/controller/RaceController.java
@@ -6,9 +6,9 @@ import lotto.model.domain.BonusNumber;
 import lotto.model.domain.Lotto;
 import lotto.model.domain.Lottos;
 import lotto.model.domain.PurchaseAmount;
-import lotto.model.domain.game.Rank;
+import lotto.model.domain.operation.Rank;
 import lotto.model.domain.WinningNumber;
-import lotto.model.domain.game.WinningStatistics;
+import lotto.model.domain.operation.WinningStatistics;
 import lotto.model.service.LottoService;
 import lotto.model.service.StatisticsService;
 import lotto.util.InputParser;
